Updated ShippingEstimateForm and added a special constraint to TableShippingMethod and ZonedShippingMethod
Used the getCountryField method of the Shop's Address model to create either a dropdown or a readonly field for the Country field in the form.

Added a special constraint to TableShippingMethod and ZonedShippingMethod, which includes rates that have no constraint values defined (e.g. min or max for any of weight, volume, value, quantity).
Added tests, and modified ZonedShippingMethodTest to share code with TableShippingMethodTest.

Updated ShippingEstimates.ss for NIL results within loop.  The results for the template are provided by the public function ShippingEstimates within CartPageShippingExtension class.  This function is called twice, once from the 'if' in the template and then 'loop'.  When called a second time, it looks for the session variable that was cleared in the initial call (that was when the 'if' was called from the template).  Then when the loop step occurs, the data has gone.  Therefore, remove the 'if' from the template and call only once from 'loop'.  Display a blank table row if there is no data.

Updated ShippingEstimateFormTests to test submitting with a dropdown field, bad data, and a readonly field.
If SS >=3.2, use SapphireTest's useTestTheme function otherwise use $this->useTheme.  These functions direct the test to a dedicated theme that includes <% include Shipping Estimator %>.